requirements.txt
absl-py==1.4.0
astor==0.8.1
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
cycler==0.11.0
deap==1.3.3
flatbuffers==23.5.9
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
joblib==1.2.0
keras==2.12.0
keras-Applications==1.0.8
keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
lightgbm==3.3.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pickleshare==0.7.5
Pillow==9.5.0
pip==23.1.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.10.1
seaborn==0.12.2
setuptools==67.7.2
six==1.16.0
stopit==1.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
TPOT==0.11.7
tqdm==4.65.0
typing_extensions==4.5.0
ultralytics==8.0.109
update-checker==0.18.0
urllib3==2.0.2
Werkzeug==2.3.4
wheel==0.40.0
wrapt==1.14.1
xgboost==1.7.5
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zipp==3.15.0
timm